+++ Src/FluentAssertions/Equivalency/EquivalencyAssertionOptions.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
 using FluentAssertions.Common;
+using FluentAssertions.Equivalency.Execution;
 using FluentAssertions.Equivalency.Ordering;
 using FluentAssertions.Equivalency.Selection;