+++ Src/FluentAssertions/EquivalencyPlan.cs
@@ -5,6 +5,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions.Equivalency;
+using FluentAssertions.Equivalency.Steps;
 
 #endregion